GPU Display: Move different source files to subfolders to clean up display folders a bit"
