+++ GPU/GPUTracking/display/backend/GPUDisplayBackend.cxx
@@ -13,7 +13,7 @@
 /// \author David Rohr
 
 #include "GPUDisplayBackend.h"
-#include "GPUDisplayMagneticField.h"
+#include "helpers/GPUDisplayMagneticField.h"
 
 #include "GPUDisplayBackendOpenGL.h"
