+++ GPU/GPUTracking/display/frontend/GPUDisplayFrontendWindows.cxx
@@ -17,7 +17,7 @@
 #include <cstdlib>
 #include <cstring>
 #include "GPUDisplayFrontendWindows.h"
-#include "GPUDisplayBackend.h"
+#include "backend/GPUDisplayBackend.h"
 #include "GPUDisplayGUIWrapper.h"
 #include "GPULogging.h"
 #include <windows.h>
